+++ src/main/java/woowahan/anifarm/tecolearning/study/controller/StudyApiController.java
@@ -8,7 +8,7 @@ import woowahan.anifarm.tecolearning.study.service.dto.StudyCreateDto;
 import woowahan.anifarm.tecolearning.study.service.dto.StudyDetailInfoDto;
 import woowahan.anifarm.tecolearning.study.service.dto.StudyInfoDto;
 import woowahan.anifarm.tecolearning.study.service.dto.StudyUpdateDto;
-import woowahan.anifarm.tecolearning.user.dto.UserInfoDto;
+import woowahan.anifarm.tecolearning.user.service.dto.UserInfoDto;
 
 @RestController
 @RequestMapping("/api/studies")
