+++ src/main/java/woowahan/anifarm/tecolearning/study/service/dto/StudyDetailInfoDto.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 import woowahan.anifarm.tecolearning.study.domain.Study;
 import woowahan.anifarm.tecolearning.studyoutput.service.dto.StudyOutputDto;
-import woowahan.anifarm.tecolearning.user.dto.UserInfoDto;
+import woowahan.anifarm.tecolearning.user.service.dto.UserInfoDto;
 
 import java.time.LocalDate;
 import java.util.List;
