+++ src/main/java/woowahan/anifarm/tecolearning/study/service/dto/StudySummaryDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 import woowahan.anifarm.tecolearning.study.domain.Study;
-import woowahan.anifarm.tecolearning.user.dto.UserInfoDto;
+import woowahan.anifarm.tecolearning.user.service.dto.UserInfoDto;
 
 @Getter
 @NoArgsConstructor
