+++ src/main/java/woowahan/anifarm/tecolearning/user/service/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package woowahan.anifarm.tecolearning.user.exception;
+package woowahan.anifarm.tecolearning.user.service.exception;
 
 public class UserNotFoundException extends RuntimeException {
     private static final String MESSAGE = "사용자를 찾지 못했습니다.";
