+++ src/test/java/woowahan/anifarm/tecolearning/study/service/StudyParticipantServiceTest.java
@@ -12,7 +12,7 @@ import woowahan.anifarm.tecolearning.study.domain.exception.PresenterException;
 import woowahan.anifarm.tecolearning.study.domain.repository.StudyParticipantRepository;
 import woowahan.anifarm.tecolearning.study.service.exception.StudyParticipantNotFoundException;
 import woowahan.anifarm.tecolearning.user.domain.User;
-import woowahan.anifarm.tecolearning.user.dto.UserInfoDto;
+import woowahan.anifarm.tecolearning.user.service.dto.UserInfoDto;
 
 import java.util.Arrays;
 import java.util.HashSet;
