+++ src/test/java/woowahan/anifarm/tecolearning/user/controller/UserApiControllerTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import woowahan.anifarm.tecolearning.AbstractWebTestClient;
 import woowahan.anifarm.tecolearning.auth.service.exception.JWTValidException;
-import woowahan.anifarm.tecolearning.user.dto.UserInfoDto;
+import woowahan.anifarm.tecolearning.user.service.dto.UserInfoDto;
 
 import java.io.UnsupportedEncodingException;
 import java.nio.charset.StandardCharsets;
